Refactor ReadWriteData Swift examples to separate CRUD pages
