+++ source/sdk/swift/crud/create.txt
@@ -96,12 +96,8 @@ the transaction immediately after the callback.
       .. tab::
          :tabid: swift
 
-         .. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.transaction.swift
+         .. literalinclude:: /examples/generated/code/start/CreateRealmObjects.snippet.transaction.swift
             :language: swift
-
-.. seealso::
-
-   :ref:`Usage Example: Read and Write Data <ios-read-and-write-data>`
 
 .. _ios-interface-driven-writes:
 .. _ios-silent-writes:
@@ -155,7 +151,7 @@ The examples on this page use the following models:
    .. tab::
       :tabid: swift
       
-      .. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.models.swift
+      .. literalinclude:: /examples/generated/code/start/CreateRealmObjects.snippet.models.swift
          :language: swift
 
    .. tab::
@@ -179,7 +175,7 @@ Create an Object
       <Structs/Realm.html#/s:10RealmSwift0A0V3add_6updateySo0aB6ObjectC_AC12UpdatePolicyOtF>`
       inside of a write transaction.
 
-      .. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.create.swift
+      .. literalinclude:: /examples/generated/code/start/CreateRealmObjects.snippet.create.swift
          :language: swift
 
    .. tab::
@@ -220,7 +216,7 @@ values:
    .. tab::
       :tabid: swift
 
-      .. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.nested-objects.swift
+      .. literalinclude:: /examples/generated/code/start/CreateRealmObjects.snippet.nested-objects.swift
          :language: swift
 
    .. tab::
@@ -256,7 +252,7 @@ convert JSON into a value that you can pass to
    .. tab::
       :tabid: swift
 
-      .. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.json.swift
+      .. literalinclude:: /examples/generated/code/start/CreateRealmObjects.snippet.json.swift
          :language: swift
 
    .. tab::
@@ -323,7 +319,7 @@ Copy an Object to Another Realm
       object to :swift-sdk:`Realm.create(_:value:update:):
       <Structs/Realm.html#/s:10RealmSwift0A0V6create_5value6updatexxm_ypAC12UpdatePolicyOtSo0aB6ObjectCRbzlF>`:
 
-      .. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.copy-to-another-realm.swift
+      .. literalinclude:: /examples/generated/code/start/CreateRealmObjects.snippet.copy-to-another-realm.swift
          :language: swift
 
    .. tab::
