+++ source/sdk/swift/crud/delete.txt
@@ -39,7 +39,7 @@ The examples on this page use the following models:
    .. tab::
       :tabid: swift
       
-      .. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.models.swift
+      .. literalinclude:: /examples/generated/code/start/DeleteRealmObjects.snippet.models.swift
          :language: swift
 
    .. tab::
@@ -61,7 +61,7 @@ Delete an Object
       <Structs/Realm.html#/s:10RealmSwift0A0V6deleteyySo13RLMObjectBaseCF>`
       inside of a write transaction.
 
-      .. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.delete.swift
+      .. literalinclude:: /examples/generated/code/start/DeleteRealmObjects.snippet.delete.swift
          :language: swift
 
    .. tab::
@@ -92,7 +92,7 @@ Delete Multiple Objects
       <Structs/Realm.html#/s:10RealmSwift0A0V6deleteyyxSTRzSo13RLMObjectBaseC7ElementRczlF>`
       inside of a write transaction.
 
-      .. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.tsq-delete-collection.swift
+      .. literalinclude:: /examples/generated/code/start/DeleteRealmObjects.snippet.tsq-delete-collection.swift
          :language: swift
 
    .. tab:: Swift NSPredicate
@@ -103,7 +103,7 @@ Delete Multiple Objects
       <Structs/Realm.html#/s:10RealmSwift0A0V6deleteyyxSTRzSo13RLMObjectBaseC7ElementRczlF>`
       inside of a write transaction.
 
-      .. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.delete-collection.swift
+      .. literalinclude:: /examples/generated/code/start/DeleteRealmObjects.snippet.delete-collection.swift
          :language: swift
 
    .. tab:: Objective C
@@ -134,7 +134,7 @@ Delete All Objects of a Specific Type
       <Structs/Realm.html#/s:10RealmSwift0A0V6deleteyyxSTRzSo13RLMObjectBaseC7ElementRczlF>`
       inside of a write transaction.
 
-      .. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.delete-all-of-class.swift
+      .. literalinclude:: /examples/generated/code/start/DeleteRealmObjects.snippet.delete-all-of-class.swift
          :language: swift
 
    .. tab::
@@ -168,7 +168,7 @@ Delete All Objects in a Realm
       write transaction. This clears the realm of all object instances
       but does not affect the realm's schema.
 
-      .. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.delete-all.swift
+      .. literalinclude:: /examples/generated/code/start/DeleteRealmObjects.snippet.delete-all.swift
          :language: swift
 
    .. tab::
