+++ source/sdk/swift/crud/threading.txt
@@ -101,7 +101,7 @@ Things to consider when performing background writes:
 - Async writes block closing or invalidating the realm
 - You can explicitly commit or cancel transactions
 
-.. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.async-transaction.swift
+.. literalinclude:: /examples/generated/code/start/UpdateRealmObjects.snippet.async-transaction.swift
    :language: swift
 
 Wait for Async Writes to Complete
