+++ source/sdk/swift/crud/update.txt
@@ -30,7 +30,7 @@ The examples on this page use the following models:
    .. tab::
       :tabid: swift
       
-      .. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.models.swift
+      .. literalinclude:: /examples/generated/code/start/UpdateRealmObjects.snippet.models.swift
          :language: swift
 
    .. tab::
@@ -51,7 +51,7 @@ Objective-C object.
    .. tab::
       :tabid: swift
       
-      .. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.update.swift
+      .. literalinclude:: /examples/generated/code/start/UpdateRealmObjects.snippet.update.swift
          :language: swift
 
    .. tab::
@@ -83,7 +83,7 @@ Applying KVC to a collection is a great way to update objects in bulk.
 Avoid the overhead of iterating over a collection while creating 
 accessors for every item.
 
-.. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.key-value-coding.swift
+.. literalinclude:: /examples/generated/code/start/UpdateRealmObjects.snippet.key-value-coding.swift
    :language: swift
 
 .. tip::
@@ -109,7 +109,7 @@ the object already exists. Upserts require the data model to have a
       <Structs/Realm.html#/s:10RealmSwift0A0V3add_6updateySo0aB6ObjectC_AC12UpdatePolicyOtF>`
       with the second parameter, update policy, set to ``.modified``.
 
-      .. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.upsert.swift
+      .. literalinclude:: /examples/generated/code/start/UpdateRealmObjects.snippet.upsert.swift
          :language: swift
 
    .. tab::
@@ -130,7 +130,7 @@ subset of the values to update:
    .. tab::
       :tabid: swift
 
-      .. literalinclude:: /examples/generated/code/start/ReadWriteData.snippet.partial-update.swift
+      .. literalinclude:: /examples/generated/code/start/UpdateRealmObjects.snippet.partial-update.swift
          :language: swift
 
    .. tab::
